requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-tenants==3.6.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2
sentry-sdk==2.13.0
python-dateutil==2.9.0.post0
gunicorn==22.0.0
pytest==7.4.0
pytest-django==4.5.2
django-filter==24.2
Pillow==10.2.0  
requests==2.31.0  

